refactor: shrink json-processing public surface (un-export internal helper, relocate import remap to app-core)
getTemplateResolvedForLegacyDataset is consumed only inside json-processing
(template-usermeta.ts internally, plus its own tests) - dropped from the
package's public index, module export kept for internal/test imports.

remapSupportFieldConfigurationForImport had no json-processing-internal
dependencies (pure function over UsermetaDatasetField/SupportFieldConfiguration
types) and its only consumer was app-core's create-button (project-create
feature). Relocated the function and its unit tests into
packages/app-core/src/lib/project, adjacent to that consumer, and updated
the create-button import. Verified against app-core's dependency graph:
data-core (source of both types) was already a dependency, so no new
dependency edge was needed.

Verified via grep across packages/*/src, src, apps, bin (excluding dist/
and node_modules) that no other consumers exist for either symbol.
architecture-boundaries canary and both packages' test suites pass.
