+++ packages/app-core/src/lib/project/index.ts
@@ -1,2 +1,5 @@
-export { isProjectInitialized } from './utils';
+export {
+    isProjectInitialized,
+    remapSupportFieldConfigurationForImport
+} from './utils';
 export type * from './types';
