+++ packages/json-processing/src/__test__/template-usermeta.test.ts
@@ -13,8 +13,7 @@ import {
     getTemplateResolvedForLegacyVersions,
     getTemplateResolvedForPlaceholderAssignment,
     getUpdatedExportMetadata,
-    getValidatedTemplate,
-    remapSupportFieldConfigurationForImport
+    getValidatedTemplate
 } from '../template-usermeta';
 import { INTERACTIVITY_DEFAULTS } from '@deneb-viz/powerbi-compat/interactivity';
 import {
@@ -1591,213 +1590,6 @@ describe('getPublishableUsermeta — supportFieldConfiguration inline in dataset
             formatted: false,
             treatAsParameter: true,
             names: true
-        });
-    });
-});
-
-describe('remapSupportFieldConfigurationForImport', () => {
-    it('should extract inline config from dataset entries keyed by suppliedObjectName', () => {
-        const dataset: UsermetaDatasetField[] = [
-            {
-                key: '__dataset.0__',
-                name: 'Date',
-                suppliedObjectName: 'Order Date',
-                kind: 'column',
-                type: 'dateTime',
-                supportFieldConfiguration: {
-                    highlight: true,
-                    highlightStatus: false,
-                    highlightComparator: false,
-                    format: false,
-                    formatted: false
-                }
-            },
-            {
-                key: '__dataset.1__',
-                name: '$ Sales',
-                suppliedObjectName: 'Revenue',
-                kind: 'measure',
-                type: 'numeric',
-                supportFieldConfiguration: {
-                    highlight: false,
-                    highlightStatus: false,
-                    highlightComparator: false,
-                    format: true,
-                    formatted: true
-                }
-            }
-        ];
-
-        const result = remapSupportFieldConfigurationForImport(dataset);
-
-        expect(result).toEqual({
-            'Order Date': {
-                highlight: true,
-                highlightStatus: false,
-                highlightComparator: false,
-                format: false,
-                formatted: false
-            },
-            Revenue: {
-                highlight: false,
-                highlightStatus: false,
-                highlightComparator: false,
-                format: true,
-                formatted: true
-            }
-        });
-    });
-
-    it('should return empty object for empty dataset', () => {
-        const result = remapSupportFieldConfigurationForImport([]);
-        expect(result).toEqual({});
-    });
-
-    it('should skip fields without supportFieldConfiguration', () => {
-        const dataset: UsermetaDatasetField[] = [
-            {
-                key: '__dataset.0__',
-                name: 'Date',
-                suppliedObjectName: 'Order Date',
-                kind: 'column',
-                type: 'dateTime'
-            }
-        ];
-
-        const result = remapSupportFieldConfigurationForImport(dataset);
-        expect(result).toEqual({});
-    });
-
-    it('should skip fields without suppliedObjectName', () => {
-        const dataset: UsermetaDatasetField[] = [
-            {
-                key: '__dataset.0__',
-                name: 'Date',
-                kind: 'column',
-                type: 'dateTime',
-                supportFieldConfiguration: {
-                    highlight: true,
-                    highlightStatus: false,
-                    highlightComparator: false,
-                    format: false,
-                    formatted: false
-                }
-            }
-        ];
-
-        const result = remapSupportFieldConfigurationForImport(dataset);
-        expect(result).toEqual({});
-    });
-
-    it('should handle treatAsParameter and names flags in roundtrip', () => {
-        // Export: config embedded per-field in dataset (via getPublishableUsermeta)
-        // Import: config extracted from dataset entries (via remapSupportFieldConfigurationForImport)
-        const supportFieldConfiguration = {
-            'Category Parameter': {
-                highlight: false,
-                highlightStatus: false,
-                highlightComparator: false,
-                format: false,
-                formatted: false,
-                treatAsParameter: true,
-                names: true
-            }
-        };
-        const exportUsermeta: UsermetaTemplate = {
-            ...EXPECTED_METADATA_BASE,
-            datasets: {
-                dataset: [
-                    {
-                        key: '__dataset.0__',
-                        name: '$ Sales',
-                        namePlaceholder: '$ Sales',
-                        description: '',
-                        kind: 'measure',
-                        type: 'numeric'
-                    },
-                    {
-                        key: '__dataset.1__',
-                        name: '$ Profit',
-                        namePlaceholder: '$ Profit',
-                        description: '',
-                        kind: 'measure',
-                        type: 'numeric'
-                    },
-                    {
-                        key: '__dataset.2__',
-                        name: 'Category Parameter',
-                        namePlaceholder: 'Category Parameter',
-                        description: '',
-                        kind: 'parameter',
-                        type: 'other'
-                    }
-                ]
-            }
-        };
-
-        const exportResult = getPublishableUsermeta(exportUsermeta, {
-            informationTranslationPlaceholders: {
-                name: '',
-                description: '',
-                author: ''
-            },
-            supportFieldConfiguration,
-            trackedFields: TRACKED_FIELDS_BY_QUERYNAME
-        });
-
-        // Export should embed config on Category Parameter (index 2)
-        expect(
-            exportResult.datasets.dataset[2].supportFieldConfiguration
-        ).toEqual({
-            highlight: false,
-            highlightStatus: false,
-            highlightComparator: false,
-            format: false,
-            formatted: false,
-            treatAsParameter: true,
-            names: true
-        });
-
-        // Now import it back with a user-supplied field name
-        const importDataset: UsermetaDatasetField[] = [
-            {
-                key: '__dataset.0__',
-                name: '$ Sales',
-                suppliedObjectName: 'Revenue',
-                kind: 'measure',
-                type: 'numeric'
-            },
-            {
-                key: '__dataset.1__',
-                name: '$ Profit',
-                suppliedObjectName: 'Gross Profit',
-                kind: 'measure',
-                type: 'numeric'
-            },
-            {
-                key: '__dataset.2__',
-                name: 'Category Parameter',
-                suppliedObjectName: 'Region Parameter',
-                kind: 'parameter',
-                type: 'other',
-                supportFieldConfiguration:
-                    exportResult.datasets.dataset[2].supportFieldConfiguration
-            }
-        ];
-
-        const importResult =
-            remapSupportFieldConfigurationForImport(importDataset);
-
-        expect(importResult).toEqual({
-            'Region Parameter': {
-                highlight: false,
-                highlightStatus: false,
-                highlightComparator: false,
-                format: false,
-                formatted: false,
-                treatAsParameter: true,
-                names: true
-            }
         });
     });
 });
